+++ src/app/api/v1/retrofunding/rounds/[roundId]/ballots/[ballotCasterAddressOrEns]/route.ts
@@ -1,0 +1,12 @@
+import { fetchBallot } from "@/app/api/common/ballots/getBallots";
+
+export async function GET(
+  request: Request,
+  route: { params: { roundId: string; ballotCasterAddressOrEns: string } }
+) {
+  const { roundId, ballotCasterAddressOrEns } = route.params;
+  const ballots = await fetchBallot(roundId, ballotCasterAddressOrEns);
+  return new Response(JSON.stringify(ballots), {
+    status: 200,
+  });
+}
