+++ src/app/api/v1/retrofunding/rounds/[roundId]/ballots/[ballotCasterAddressOrEns]/submit/route.ts
@@ -1,0 +1,12 @@
+import { fetchBallot } from "@/app/api/common/ballots/getBallots";
+
+export async function POST(
+  request: Request,
+  route: { params: { roundId: string; ballotCasterAddressOrEns: string } }
+) {
+  const { roundId, ballotCasterAddressOrEns } = route.params;
+  const ballot = await fetchBallot(roundId, ballotCasterAddressOrEns);
+  return new Response(JSON.stringify(ballot), {
+    status: 200,
+  });
+}
